Added normally distributed mutations
Also fixed a bug that caused the random weights to never be negative. Since the mutation was also always positive, weights could never evolve to a negative value either, nor get small as the mutation factor was at least 2.
The new implementation fixes both these issues, and gives room for more variety in the mutations as they're not fixed to an integer value.