Moved AI controls to FixedUpdate
The AI controller now processes its inputs on FixedUpdate instead of Update, to maintain a consistent output throughout the training, and when using the weights on a test drone later.
Before, differences in frame rate would affect the output and performance of the drones, which was most noticeable when testing the weights on an isolated drone. Since the framerate was much higher, it performed very differently compared to the training drones.
Now, the same network will have mostly the same outcomes despite framerate differences.
This also means that it is less likely, although still possible, for the best score in a generation to be worse than that of the previous one, since more of the parameters are kept unchanged (the best drone in a generation is always maintained, but there were still differences in the outcomes partly because of fluctuations in the framerate).