Changed AI inputs
The networks now receive inputs that are relative to both the drone's position and rotation, so they can be used regardless of the global position or orientation of the checkpoints.
They are now also aware of the drone's angular velocity, as well as the scalar distance to the next checkpoints.
Adjusted the network size to accommodate these new inputs.